+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Added `upgradePackages` directive for the configuration file. This allows to
+  update packages in the base image.
+
 ## [0.14.0] - 2025-03-21
 
 ### Added